+++ web-registry/src/lib/queries/react-query/marketplace/getSellOrdersExtendedQuery/getSellOrdersExtendedQuery.types.ts
